rcbd/requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
bert-score>=0.3.13
statsmodels>=0.14.0
matplotlib>=3.5.0
seaborn>=0.11.0
torch>=2.0.0
transformers>=4.21.0
python-dotenv>=0.19.0
scikit-learn>=1.1.0
tqdm>=4.64.0
tokenizers>=0.13.0 